+++ src/Concerto/PanelBundle/Command/ConcertoRCacheCommand.php
@@ -39,12 +39,13 @@ class ConcertoRCacheCommand extends Command
         $p = "\"" . $this->testRunnerSettings['rscript_exec'] . "\" --no-save --no-restore --quiet \"" . $script_path . "\"";
         $process = new Process($p);
 
-        $r_environ_path = $this->testRunnerSettings['r_environ_path'];
-        if ($r_environ_path != null) {
-            $env = array();
-            $env["R_ENVIRON"] = $r_environ_path;
-            $process->setEnv($env);
-        }
+        $r_environ_path = $this->testRunnerSettings['r_environ_session_path'];
+        $r_profile_path = $this->testRunnerSettings['r_profile_session_path'];
+        $env = [
+            "R_ENVIRON_USER" => $r_environ_path != null ? $r_environ_path : "{$this->projectDir}/app/config/R/.Renviron_session",
+            "R_PROFILE_USER" => $r_profile_path != null ? $r_profile_path : "{$this->projectDir}/app/config/R/.Rprofile_session"
+        ];
+        $process->setEnv($env);
 
         $process->setTimeout(null);
         $process->run();