+++ src/Concerto/TestBundle/Command/StartForkerCommand.php
@@ -64,6 +64,8 @@ class StartForkerCommand extends Command
         $process = new Process($cmd);
         $process->inheritEnvironmentVariables(true);
 
+        $r_environ_path = $this->testRunnerSettings["r_environ_session_path"];
+        $r_profile_path = $this->testRunnerSettings['r_profile_session_path'];
         $env = [
             "CONCERTO_R_APP_URL" => $appUrl,
             "CONCERTO_R_DB_CONNECTION" => $dbConnection,
@@ -77,11 +79,12 @@ class StartForkerCommand extends Command
             "CONCERTO_R_SESSION_STORAGE" => $sessionStorage,
             "CONCERTO_R_SESSION_FILES_EXPIRATION" => $sessionFilesExpiration,
             "CONCERTO_R_SESSION_LOG_LEVEL" => $sessionLogLevel,
-            "R_GC_MEM_GROW" => 0
+            "R_GC_MEM_GROW" => 0,
+            "R_ENVIRON_USER" => $r_environ_path != null ? $r_environ_path : "{$this->projectDir}/app/config/R/.Renviron_session",
+            "R_PROFILE_USER" => $r_profile_path != null ? $r_profile_path : "{$this->projectDir}/app/config/R/.Rprofile_session"
         ];
-        if ($this->testRunnerSettings["r_environ_path"] != null) $env["R_ENVIRON"] = $this->testRunnerSettings["r_environ_path"];
-
         $process->setEnv($env);
+
         $process->mustRun();
         if ($process->getExitCode() == 0) {
             $output->writeln("forker started");